Option stack is resolved "bottom-up" to abandonment
The strategy to abandon an option is "bottom-up" in relation to
salient events:

The function goes through the option stack from the oldest one to
the newest one (from the bottom to the top).  If the current state
is a salient event to an option, this option is removed from the
stack, along with all the other options (if any) that are newer
than this one.